Refactor all web components to follow Atomic Design principles with
Tb-prefix naming convention for better maintainability and consistency.

Features:
- Atomic Design hierarchy (Atoms → Molecules → Organisms → Features)
- Tb-prefix for all custom components (TbButton, TbCard, etc.)
- Feature-based component organization (Dashboard, Portfolio, Strategy, etc.)
- Consolidated imports with proper hierarchy
- Improved component reusability and testability
- Enhanced component documentation
- Consistent prop interfaces across components
- Better separation of concerns
- Upgrade to .NET 10 with latest dependencies
- Updated Spectre.Console packages to 0.54.0

+++ tests/TradingBot.Analytics.Tests/PerformanceCalculatorTests.cs
@@ -2,8 +2,6 @@
 // Copyright (c) TradingBot. All rights reserved.
 // </copyright>
 
-using Shouldly;
-using TradingBot.Analytics;
 using TradingBot.Core.Enums;
 using TradingBot.Core.Models.Trading;
 
@@ -346,7 +344,7 @@ public sealed class PerformanceCalculatorTests
 
         for (int i = 0; i < count; i++)
         {
-            var entryPrice = 100m + (decimal)random.Next(-20, 20);
+            var entryPrice = 100m + random.Next(-20, 20);
             var change = (decimal)(random.NextDouble() * 20) - 10m; // -10% to +10%
             var exitPrice = entryPrice * (1m + (change / 100m));
